requirements.txt
backcall==0.2.0
certifi==2020.12.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.2
Django==2.2.17
djangorestframework==3.12.2
idna==2.10
ipython==7.9.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.17.2
Jinja2==2.11.2
MarkupSafe==1.1.1
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==2.0.10
psycopg2==2.8.6
ptyprocess==0.6.0
Pygments==2.7.2
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.26.2
wcwidth==0.2.5
